feat: correct action names, wrap f in useCallback
