+++ lessons/hw44-todo/package.json
@@ -17,6 +17,7 @@
     "react-redux": "^7.2.6",
     "react-scripts": "4.0.3",
     "redux": "^4.1.2",
+    "redux-devtools-extension": "^2.13.9",
     "typescript": "^4.4.4",
     "uuid": "^8.3.2",
     "web-vitals": "^1.1.2"
