+++ lessons/hw44-todo/src/App.css
@@ -13,6 +13,7 @@
   flex-direction: column;
   gap: 10px;
   border: 1px solid lightgray;
+  box-shadow: 5px 5px 15px lightgray;
 }
