+++ lessons/hw44-todo/src/store/index.ts
@@ -1,6 +1,6 @@
 import { createStore, combineReducers } from 'redux'
 import { tasksReducer } from './tasksReducer';
-// import { todoListReducer } from './todoListReducer';
+import { composeWithDevTools } from 'redux-devtools-extension';
 
 const rootReducer = combineReducers({
     tasks: tasksReducer
@@ -11,4 +11,5 @@ export type RootState = ReturnType<typeof rootReducer>
 // export const store = createStore(rootReducer, (window as any).__REDUX_DEVTOOLS_EXTENSION__ &&
 // (window as any).__REDUX_DEVTOOLS_EXTENSION__());
 
-export const store = createStore(rootReducer, (window as any).__REDUX_DEVTOOLS_EXTENSION__ && (window as any).__REDUX_DEVTOOLS_EXTENSION__());
+// export const store = createStore(rootReducer, (window as any).__REDUX_DEVTOOLS_EXTENSION__ && (window as any).__REDUX_DEVTOOLS_EXTENSION__());
+export const store = createStore(rootReducer, composeWithDevTools());
